Version_10
the basic game is finished.
It still is with the buttons which I don't want, but everything else game wise is done.
I intend to go back and make it look nicer and all the different scripts could be neater and checked over again for leftover junk from testing.
but besides all that it's playable!